+++ include/LayerCodeInCellID.h
@@ -11,10 +11,10 @@ class LayerCodeInCellID {
     virtual ~LayerCodeInCellID() {}
 
     int GetLayer(std::string const CellID, int const StartBin_layers, int const LengthBin_layers);
-    int GetLayer(unsigned long const cellid, int const StartBin_layers, int const LengthBin_layers);
+    int GetLayer(unsigned long long const cellid, int const StartBin_layers, int const LengthBin_layers);
 
   private:
-    unsigned long Layer;
+    unsigned long long Layer;
 
     int FindLayer(std::string const CellID_, int const StartBin_layers, int const LengthBin_layers);
 };
